chore: move registerGuest to omni-core
