+++ apps/meteor/app/livechat/imports/server/rest/sms.ts
@@ -10,6 +10,7 @@ import type {
 import { OmnichannelSourceType } from '@rocket.chat/core-typings';
 import { Logger } from '@rocket.chat/logger';
 import { LivechatVisitors, LivechatRooms, LivechatDepartment } from '@rocket.chat/models';
+import { registerGuest } from '@rocket.chat/omni-core';
 import { Random } from '@rocket.chat/random';
 import { serverFetch as fetch } from '@rocket.chat/server-fetch';
 import { Meteor } from 'meteor/meteor';
@@ -20,7 +21,6 @@ import { FileUpload } from '../../../../file-upload/server';
 import { checkUrlForSsrf } from '../../../../lib/server/functions/checkUrlForSsrf';
 import { settings } from '../../../../settings/server';
 import { setCustomField } from '../../../server/api/lib/customFields';
-import { registerGuest } from '../../../server/lib/guests';
 import type { ILivechatMessage } from '../../../server/lib/localTypes';
 import { sendMessage } from '../../../server/lib/messages';
 import { createRoom } from '../../../server/lib/rooms';
@@ -59,8 +59,9 @@ const defineDepartment = async (idOrName?: string) => {
 
 const defineVisitor = async (smsNumber: string, targetDepartment?: string) => {
 	const visitor = await LivechatVisitors.findOneVisitorByPhone(smsNumber);
-	let data: { token: string; department?: string } = {
+	let data: { token: string; department?: string; shouldConsiderIdleAgent: boolean } = {
 		token: visitor?.token || Random.id(),
+		shouldConsiderIdleAgent: settings.get<boolean>('Livechat_enabled_when_agent_idle'),
 	};
 
 	if (!visitor) {
