+++ packages/omni-core/src/index.ts
@@ -1,1 +1,2 @@
 export * from './isDepartmentCreationAvailable';
+export * from './visitor/create';
